feat: de-experimentalize pipe API and verify depth ceilings
Replace the recursion-warning notes on the composition API with measured
per-library ceilings. The matrix test exercises each endorsed pipe
library at the deepest pipeline whose result still resolves through
FromSchema: pipe-ts @ 9 (overload cap), remeda @ 15 (overload cap),
effect @ 15 and ts-functional-pipe @ 16 (TS2589 ceiling).
