+++ test/composition-pipe-depth-ceiling-mock.ts
